simplified clients to one object
