+++ ynabsplitbudget/client.py
@@ -5,24 +5,50 @@ from typing import List, Union
 import requests
 from requests import HTTPError
 
+from ynabsplitbudget.models.account import Account
+from ynabsplitbudget.models.exception import BudgetNotFound, AccountNotFound
 from ynabsplitbudget.transactionbuilder import TransactionBuilder
 from ynabsplitbudget.models.transaction import RootTransaction, LookupTransaction, ComplementTransaction
-from ynabsplitbudget.models.user import User
 
 YNAB_BASE_URL = 'https://api.ynab.com/v1/'
 
 
 @dataclass
-class SyncClient:
+class Client:
 
-	def __init__(self, user: User):
+	def __init__(self, user_name: str, budget_id: str, account_id: str, token: str):
 		self.session = requests.Session()
-		self.session.headers.update({'Authorization': f'Bearer {user.token}'})
-		self.user = user
-		self.transaction_builder = TransactionBuilder(user=user)
+		self.session.headers.update({'Authorization': f'Bearer {token}'})
+		self.user_name = user_name
+		self.budget_id = budget_id
+		self.account_id = account_id
+		self.transaction_builder = TransactionBuilder(account_id=self.account_id)
+
+	def fetch_account(self, budget_id: str, account_id: str) -> Account:
+		r = self.session.get(f'{YNAB_BASE_URL}budgets', params=dict(include_accounts=True))
+		r.raise_for_status()
+		data_dict = r.json()['data']
+
+		try:
+			budget = next(b for b in data_dict['budgets'] if b['id'] == budget_id)
+		except StopIteration:
+			raise BudgetNotFound(f"No budget with id '{budget_id} found for {self.user_name}'")
+
+		try:
+			account = next(a for a in budget['accounts'] if a['id'] == account_id and a['deleted'] is False)
+		except StopIteration:
+			raise AccountNotFound(f"No Account with id '{account_id}' fund in budget '{budget['name']} "
+								  f"for user {self.user_name}'")
+
+		return Account(budget_id=budget_id,
+					   budget_name=budget['name'],
+					   account_id=account_id,
+					   account_name=account['name'],
+					   transfer_payee_id=account['transfer_payee_id'],
+					   currency=budget['currency_format']['iso_code'])
 
 	def fetch_roots(self, since: date) -> List[RootTransaction]:
-		url = f'{YNAB_BASE_URL}budgets/{self.user.budget_id}/accounts/{self.user.account_id}/transactions'
+		url = f'{YNAB_BASE_URL}budgets/{self.budget_id}/accounts/{self.account_id}/transactions'
 		r = self.session.get(url, params={'since_date': datetime.strftime(since, '%Y-%m-%d')})
 		r.raise_for_status()
 		transactions_dicts = r.json()['data']['transactions']
@@ -34,7 +60,7 @@ class SyncClient:
 		return transactions
 
 	def fetch_lookup(self, since: date) -> List[Union[RootTransaction, LookupTransaction, ComplementTransaction]]:
-		url = f'{YNAB_BASE_URL}budgets/{self.user.budget_id}/transactions'
+		url = f'{YNAB_BASE_URL}budgets/{self.budget_id}/transactions'
 		r = self.session.get(url, params={'since_date': datetime.strftime(since, '%Y-%m-%d')})
 		r.raise_for_status()
 		data_dict = r.json()['data']['transactions']
@@ -53,9 +79,9 @@ class SyncClient:
 					iteration += 1
 
 	def _insert(self, t: RootTransaction, iteration: int):
-		url = f'{YNAB_BASE_URL}budgets/{self.user.budget_id}/transactions'
+		url = f'{YNAB_BASE_URL}budgets/{self.budget_id}/transactions'
 		data = {'transaction': {
-			"account_id": self.user.account_id,
+			"account_id": self.account_id,
 			"date": t.transaction_date.strftime("%Y-%m-%d"),
 			"amount": - t.amount,
 			"payee_name": t.payee_name,
@@ -68,13 +94,13 @@ class SyncClient:
 		r.raise_for_status()
 
 	def fetch_balance(self) -> int:
-		url = f'{YNAB_BASE_URL}budgets/{self.user.budget_id}/accounts/{self.user.account_id}'
+		url = f'{YNAB_BASE_URL}budgets/{self.budget_id}/accounts/{self.account_id}'
 		r = self.session.get(url)
 		r.raise_for_status()
 		balance = r.json()['data']['account']['cleared_balance']
 		return balance
 
 	def delete_complement(self, transaction_id: str) -> None:
-		url = f'{YNAB_BASE_URL}budgets/{self.user.budget_id}/transactions/{transaction_id}'
+		url = f'{YNAB_BASE_URL}budgets/{self.budget_id}/transactions/{transaction_id}'
 		r = self.session.delete(url)
 		r.raise_for_status()
